requirements.txt
asgiref==3.8.1
Django==5.1.3
django-seed==0.3.1
djangorestframework==3.15.2
drf-yasg==1.21.8
Faker==33.0.0
inflection==0.5.1
numpy==2.1.3
packaging==24.2
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
six==1.16.0
sqlparse==0.5.1
toposort==1.10
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
